requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Werkzeug==3.0.3
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
bcrypt==4.1.2
PyJWT==2.8.0
cryptography>=41.0.8
gunicorn==21.2.0
psycopg2-binary
python-dotenv==1.0.0
